refactor: move rpc/tests to a separate file
